Remove `busted` testing facilities for `LuaUnit`
`busted` gives difficulties when using the `LuaJIT` interpreter--which
is what the target game framework, LOVE, uses. `LuaUnit`, however, gives
no issues and works without a hitch. Additionally, `LuaUnit` is
contained in a single file; testing requires only the `LuaJIT`
interpreter. The only loss, in my eyes, is the maturity of `busted` and
colored text in the terminal. In return, we get to actually test our ECS
in the target interpreter.