Fix comment documentation
On another note, encountered a bug(?) with constructing a filter using
`Filter()` syntax. It's been a few months since I've worked closely with
Lua, so I am unsure if this is an expected caveat of that `Filter()`
syntax or if it is an actual bug. Reading how this syntax is implemented
through metatables tells me it is a bug.